Add golden test for Cobertura report
Add a test comparing the output report of project1 in Cobertura
format. The test compares generated XML and the contents of
"test/data/golden/project1.xml".

Under Windows, using a non-existing file name to write a file to get
rid of path separator and newline differences. The test will write a
file if the golden file is missing.