Parse timestamp in .mix file
Modify "readMix'" to parse the timestamp in the .mix file. Use the
parsed timestamp instead of the dummy value in the returned "Mix"
value. Update the documentation comment of "readMix'". Remove the note
about the dummy value.

Add some tests for "readTix'" and "readMix'" parsers. Add a malformed
.mix file containing an invalid timestamp. Add the "hpc" package to
test dependency in cabal configuration.

Update the changelog. Mention the modification in "readMix'".